store 역할별로 분할하기(원래 index.js 하나의 파일안에 모든 기능들이 나열되어 있었는데, actions.js, getters.js, index.js, mutations.js, state.js 다섯개의 파로 분리)
